Moved SWE stuff to swe/ folder.
