+++ src/rdysetup.c
@@ -4,7 +4,7 @@
 #include <private/rdymemoryimpl.h>
 #include <rdycore.h>
 
-#include "swe_setup_ceed.h"
+#include "swe/swe_setup_ceed.h"
 
 // Maximum length of the name of a prognostic or diagnostic field component
 #define MAX_COMP_NAME_LENGTH 20
