+++ src/swe/swe_setup_ceed.h
